Testify: Add Struct & Function for HTML Replacement
* add a testify.ReplaceHTMLContent function that takes in web
content by reference and then replaces it. This replacement is
done using a loop, which utilizes the HTMLReplacement struct by
iterating over an array of []HTMLReplacement, using the
HTMLReplacement.ReplaceString as the string to be replaced in the
web content, and HTMLReplacement.ReplacementString as the string
that the selection will be replaced with.